Remove DataSource, move files up from parsers directory
