+++ onair/data_handling/csv_parser.py
@@ -15,7 +15,7 @@ import os
 import pandas as pd
 
 from .on_air_data_source import OnAirDataSource
-from ...src.util.print_io import *
+from ..src.util.print_io import *
 from .parser_util import * 
 
 class CSV(OnAirDataSource):
